Quick Start

Quick Start

Initialising the model object

The first step is to initialise the SDDP model object. We do this using the following syntax:

If we have more than one markov state:

m = SDDPModel([;kwargs...]) do sp, stage, markov_state

    # Stage problem definition where `sp` is a `JuMP.Model`object,

end

Otherwise if we have a single markov state

m = SDDPModel([;kwargs...]) do sp, stage

  # Stage problem definition

end

SDDP.SDDPModel — Type.

SDDPModel(;kwargs...) do ...

end

Description

This function constructs an SDDPModel.

Required Keyword arguments

stages::Int

The number of stages in the problem. A stage is defined as each step in time at which a decion can be made. Defaults to 1.

objective_bound::Float64

solver::MathProgBase.AbstractMathProgSolver

Optional Keyword arguments

cut_oracle

risk_measure

noise_probability

markov_transition

Returns

* `m`: the `SDDPModel`

source

Solve

To solve the SDDP model m we use status = solve(m::SDDPModel; kwargs...). This accepts a number of keyword arguments to control the solution process.

JuMP.solve — Function.

solve(m::SDDPModel; kwargs...)

Description

Solve the SDDPModel m using SDDP. Accepts a number of keyword arguments to control the solution process.

Positional arguments

m: the SDDPModel to solve

Keyword arguments

max_iterations::Int: The maximum number of cuts to add to a single stage problem before terminating. Defaults to 10.

time_limit::Real: The maximum number of seconds (in real time) to compute for before termination. Defaults to Inf.

simulation::MonteCarloSimulation: see MonteCarloSimulation

bound_convergence::BoundConvergence: see BoundConvergence

cut_selection_frequency::Int: Frequency (by iteration) with which to rebuild subproblems using a subset of cuts. Frequent cut selection (i.e. cut_selection_frequency is small) reduces the size of the subproblems that are solved, but incurrs the overhead of rebuilding the subproblems. However, infrequent cut selection (i.e. cut_selection_frequency is large) allows the subproblems to grow large (many constraints) leading to an increase in the solve time of individual subproblems. Defaults to 0 (never run).

print_level::Int: 0 - off: nothing logged to screen (still written to log file if specified). 1 - on: solve iterations written to screen. Defaults to 1

log_file::String: Relative filename to write the log to disk. Defaults to "" (no log written)

solve_type: One of

Asyncronous() - solve using a parallelised algorithm

Serial() - solve using a serial algorithm

Default chooses automatically based on the number of available processors.

reduce_memory_footprint::Bool: Implements the idea proposed in https://github.com/JuliaOpt/JuMP.jl/issues/969#issuecomment-282191105 to reduce the memory consumption when running SDDP. This is an issue if you wish to save the model m to disk since it discards important information. Defaults to false.

cut_output_file::String: Relative filename to write discovered cuts to disk. Defaults to "" (no cuts written)

Returns

status::Symbol: Reason for termination. One of

:solving

:interrupted

:converged

:max_iterations

:bound_convergence

:time_limit

source

Simulate

Visualise

SDDP.@visualise — Macro.

@visualise(results, replication, stage, begin
	... plot definitions ...
end)

Description

Plot everything using interactive javascript. This will launch an HTML page
to explore.

Usage

    `@visualise(results, i, t, begin
        ... one line for each plot ...
    end)`

where results is the vector of result dictionaries from simulate(), i is the
simulation index (1:length(results)), and t is the stage index (1:T).

Each plot line gets transformed into an anonymous function
    (results, i, t) -> ... plot line ...
so can be any valid Julia syntax that uses results, i, or t as an argument.

After the plot definition, keyword arguments can be used (in parenthesises):
    `title`       - set the title of the plot
    `ylabel`      - set the yaxis label
    `xlabel`      - set the xaxis label
    `interpolate` - interpolate lines between stages. Defaults to "linear"
    see https://github.com/d3/d3-3.x-api-reference/blob/master/SVG-Shapes.md
        #line_interpolate for all options

Results Object

`results::Vector{Dict{Symbol, Any}}` is a vector of dictionaries where each
dictionary corresponds to one simulation (therefore there will be
`N = length(results)` lines plotted in each graph).

source