API Reference

Policy graphs

SDDP.Graph — Type

Graph(root_node::T) where T

Create an empty graph struture with the root node root_node.

SDDP.add_node — Function

add_node(graph::Graph{T}, node::T) where T

Add a node to the graph graph.

Examples

add_node(graph, :A)

SDDP.add_edge — Function

add_edge(graph::Graph{T}, edge::Pair{T, T}, probability::Float64) where T

Add an edge to the graph graph.

Examples

add_edge(graph, 1 => 2, 0.9)
add_edge(graph, :root => :A, 1.0)

SDDP.add_ambiguity_set — Function

add_ambiguity_set(graph::Graph{T}, set::Vector{T}, lipschitz::Vector{Float64})

Add set to the belief partition of graph.

lipschitz is a vector of Lipschitz constants, with one element for each node in set. The Lipschitz constant is the maximum slope of the cost-to-go function with respect to the belief state associated with each node at any point in the state-space.

Examples

graph = LinearGraph(3)
add_ambiguity_set(graph, [1, 2], [1e3, 1e2])
add_ambiguity_set(graph, [3], [1e5])

add_ambiguity_set(graph::Graph{T}, set::Vector{T}, lipschitz::Float64)

Add set to the belief partition of graph.

lipschitz is a Lipschitz constant for each node in set. The Lipschitz constant is the maximum slope of the cost-to-go function with respect to the belief state associated with each node at any point in the state-space.

Examples

graph = LinearGraph(3)
add_ambiguity_set(graph, [1, 2], 1e3)
add_ambiguity_set(graph, [3], 1e5)

SDDP.LinearGraph — Function

LinearGraph(stages::Int)

SDDP.MarkovianGraph — Function

MarkovianGraph(transition_matrices::Vector{Matrix{Float64}})

Construct a Markovian graph from the vector of transition matrices.

transition_matrices[t][i, j] gives the probability of transitioning from Markov state i in stage t - 1 to Markov state j in stage t.

The dimension of the first transition matrix should be (1, N), and transition_matrics[1][1, i] is the probability of transitioning from the root node to the Markov state i.

MarkovianGraph(;
    stages::Int,
    transition_matrix::Matrix{Float64},
    root_node_transition::Vector{Float64}
)

Construct a Markovian graph object with stages number of stages and time-independent Markov transition probabilities.

transition_matrix must be a square matrix, and the probability of transitioning from Markov state i in stage t to Markov state j in stage t + 1 is given by transition_matrix[i, j].

root_node_transition[i] is the probability of transitioning from the root node to Markov state i in the first stage.

MarkovianGraph(
    simulator::Function; budget::Union{Int, Vector{Int}}, scenarios::Int = 1000
)

Construct a Markovian graph by fitting Markov chain to scenarios generated by simulator().

budget is the total number of nodes in the resulting Markov chain. This can either be specified as a single Int, in which case we will attempt to intelligently distributed the nodes between stages. Alternatively, budget can be a Vector{Int}, which details the number of Markov state to have in each stage.

SDDP.LinearPolicyGraph — Function

LinearPolicyGraph(builder::Function; stages::Int, kwargs...)

Create a linear policy graph with stages number of stages.

See SDDP.PolicyGraph for the other keyword arguments.

SDDP.MarkovianPolicyGraph — Function

MarkovianPolicyGraph(
    builder::Function;
    transition_matrices::Vector{Array{Float64, 2}},
    kwargs...
)

Create a Markovian policy graph based on the transition matrices given in transition_matrices.

transition_matrices[t][i, j] gives the probability of transitioning from Markov state i in stage t - 1 to Markov state j in stage t.

The dimension of the first transition matrix should be (1, N), and transition_matrics[1][1, i] is the probability of transitioning from the root node to the Markov state i.

See SDDP.MarkovianGraph for other ways of specifying a Markovian policy graph. See SDDP.PolicyGraph for the other keyword arguments.

SDDP.PolicyGraph — Type

PolicyGraph(
    builder::Function,
    graph::Graph{T};
    sense::Symbol = :Min,
    lower_bound = -Inf,
    upper_bound = Inf,
    optimizer = nothing,
    bellman_function = nothing,
    direct_mode::Bool = false,
    integrality_handler = ContinuousRelaxation(),
) where {T}

Construct a policy graph based on the graph structure of graph. (See SDDP.Graph for details.)

Example

function builder(subproblem::JuMP.Model, index)
    # ... subproblem definition ...
end

model = PolicyGraph(
    builder,
    graph;
    lower_bound = 0.0,
    optimizer = GLPK.Optimizer,
    direct_mode = false
)

Or, using the Julia do ... end syntax:

model = PolicyGraph(
    graph;
    lower_bound = 0.0,
    optimizer = GLPK.Optimizer,
    direct_mode = true
) do subproblem, index
    # ... subproblem definitions ...
end

Subproblem definition

SDDP.@stageobjective — Macro

@stageobjective(subproblem, expr)

Set the stage-objective of subproblem to expr.

Example

@stageobjective(subproblem, 2x + y)

SDDP.parameterize — Function

parameterize(modify::Function,
             subproblem::JuMP.Model,
             realizations::Vector{T},
             probability::Vector{Float64} = fill(1.0 / length(realizations))
                 ) where T

Add a parameterization function modify to subproblem. The modify function takes one argument and modifies subproblem based on the realization of the noise sampled from realizations with corresponding probabilities probability.

In order to conduct an out-of-sample simulation, modify should accept arguments that are not in realizations (but still of type T).

Example

SDDP.parameterize(subproblem, [1, 2, 3], [0.4, 0.3, 0.3]) do ω
    JuMP.set_upper_bound(x, ω)
end

parameterize(node::Node, noise)

Parameterize node node with the noise noise.

SDDP.add_objective_state — Function

add_objective_state(update::Function, subproblem::JuMP.Model; kwargs...)

Add an objective state variable to subproblem.

Required kwargs are:

initial_value: The initial value of the objective state variable at the root node.

lipschitz: The lipschitz constant of the objective state variable.

Setting a tight value for the lipschitz constant can significantly improve the speed of convergence.

Optional kwargs are:

lower_bound: A valid lower bound for the objective state variable. Can be -Inf.

upper_bound: A valid upper bound for the objective state variable. Can be +Inf.

Setting tight values for these optional variables can significantly improve the speed of convergence.

If the objective state is N-dimensional, each keyword argument must be an NTuple{N, Float64}. For example, initial_value = (0.0, 1.0).

SDDP.objective_state — Function

objective_state(subproblem::JuMP.Model)

Return the current objective state of the problem.

Can only be called from SDDP.parameterize.

SDDP.Noise — Type

Noise(support, probability)

An atom of a discrete random variable at the point of support support and associated probability probability.

Training the policy

SDDP.numerical_stability_report — Function

numerical_stability_report([io::IO=stdout,] model::PolicyGraph,
                           by_node::Bool=false, print=true, warn::Bool=true)

Print a report identifying possible numeric stability issues.

If by_node, print a report for each node in the graph.

If print, print to io.

If warn, warn if the coefficients may cause numerical issues.

SDDP.train — Function

SDDP.train(model::PolicyGraph; kwargs...)

Train the policy for model. Keyword arguments:

iteration_limit::Int: number of iterations to conduct before termination.

time_limit::Float64: number of seconds to train before termination.

stoping_rules: a vector of SDDP.AbstractStoppingRules.

print_level::Int: control the level of printing to the screen. Defaults to 1. Set to 0 to disable all printing.

log_file::String: filepath at which to write a log of the training progress. Defaults to SDDP.log.

log_frequency::Int: control the frequency with which the logging is outputted (iterations/log). Defaults to 1.

run_numerical_stability_report::Bool: generate (and print) a numerical stability report prior to solve. Defaults to true.

refine_at_similar_nodes::Bool: if SDDP can detect that two nodes have the same children, it can cheaply add a cut discovered at one to the other. In almost all cases this should be set to true.

cut_deletion_minimum::Int: the minimum number of cuts to cache before deleting cuts from the subproblem. The impact on performance is solver specific; however, smaller values result in smaller subproblems (and therefore quicker solves), at the expense of more time spent performing cut selection.

risk_measure: the risk measure to use at each node. Defaults to Expectation.

sampling_scheme: a sampling scheme to use on the forward pass of the algorithm. Defaults to InSampleMonteCarlo.

backward_sampling_scheme: a backward pass sampling scheme to use on the backward pass of the algorithm. Defaults to CompleteSampler.

cut_type: choose between SDDP.SINGLE_CUT and SDDP.MULTI_CUT versions of SDDP.

dashboard::Bool: open a visualization of the training over time. Defaults to false.

parallel_scheme::AbstractParallelScheme: specify a scheme for solving in parallel. Defaults to Serial().

forwad_pass::AbstractForwardPass: specify a scheme to use for the forward passes.

There is also a special option for infinite horizon problems

cycle_discretization_delta: the maximum distance between states allowed on the forward pass. This is for advanced users only and needs to be used in conjunction with a different sampling_scheme.

SDDP.termination_status — Function

termination_status(model::PolicyGraph)

Query the reason why the training stopped.

SDDP.write_cuts_to_file — Function

write_cuts_to_file(model::PolicyGraph{T}, filename::String) where {T}

Write the cuts that form the policy in model to filename in JSON format.

See also SDDP.read_cuts_from_file.

SDDP.read_cuts_from_file — Function

read_cuts_from_file(
    model::PolicyGraph{T}, filename::String;
    node_name_parser::Function = _node_name_parser) where {T}

Read cuts (saved using SDDP.write_cuts_to_file) from filename into model.

Since T can be an arbitrary Julia type, the conversion to JSON is lossy. When reading, read_cuts_from_file only supports T=Int, T=NTuple{N, Int}, and T=Symbol. If you have manually created a policy graph with a different node type T, provide a function node_name_parser with the signature node_name_parser(T, name::String)::T where {T} that returns the name of each node given the string name name.

See also SDDP.write_cuts_to_file.

SDDP.write_log_to_csv — Function

write_log_to_csv(model::PolicyGraph, filename::String)

Write the log of the most recent training to a csv for post-analysis.

Assumes that the model has been trained via SDDP.train.

Stopping rules

SDDP.AbstractStoppingRule — Type

AbstractStoppingRule

The abstract type for the stopping-rule interface.

You need to define the following methods:

SDDP.stopping_rule_status

SDDP.convergence_test

SDDP.stopping_rule_status — Function

stopping_rule_status(::AbstractStoppingRule)::Symbol

Return a symbol describing the stopping rule.

SDDP.convergence_test — Function

convergence_test(model::PolicyGraph, log::Vector{Log}, ::AbstractStoppingRule)::Bool

Return a Bool indicating if the algorithm should terminate the training.

Sampling schemes

SDDP.AbstractSamplingScheme — Type

AbstractSamplingScheme

The abstract type for the sampling-scheme interface.

You need to define the following methods:

SDDP.sample_scenario

SDDP.sample_scenario — Function

sample_scenario(graph::PolicyGraph{T}, ::AbstractSamplingScheme) where T

Sample a scenario from the policy graph graph based on the sampling scheme.

Returns ::Tuple{Vector{Tuple{T, <:Any}}, Bool}, where the first element is the scenario, and the second element is a Boolean flag indicating if the scenario was terminated due to the detection of a cycle.

The scenario is a list of tuples (type Vector{Tuple{T, <:Any}}) where the first component of each tuple is the index of the node, and the second component is the stagewise-independent noise term observed in that node.

SDDP.InSampleMonteCarlo — Type

InSampleMonteCarlo(;
    max_depth::Int = 0,
    terminate_on_cycle::Function = false,
    terminate_on_dummy_leaf::Function = true,
    rollout_limit::Function = (i::Int) -> typemax(Int)
)

A Monte Carlo sampling scheme using the in-sample data from the policy graph definition.

If terminate_on_cycle, terminate the forward pass once a cycle is detected. If max_depth > 0, return once max_depth nodes have been sampled. If terminate_on_dummy_leaf, terminate the forward pass with 1 - probability of sampling a child node.

Note that if terminate_on_cycle = false and terminate_on_dummy_leaf = false then max_depth must be set > 0.

You can use rollout_limit to set iteration specific depth limits. For example:

InSampleMonteCarlo(rollout_limit = i -> 2 * i)

SDDP.OutOfSampleMonteCarlo — Type

OutOfSampleMonteCarlo(
    f::Function, graph::PolicyGraph;
    use_insample_transition::Bool = false,
    max_depth::Int = 0,
    terminate_on_cycle::Bool = false,
    terminate_on_dummy_leaf::Bool = true,
    rollout_limit::Function = i -> typemax(Int),
)

Create a Monte Carlo sampler using out-of-sample probabilities and/or supports for the stagewise-independent noise terms, and out-of-sample probabilities for the node-transition matrix.

f is a function that takes the name of a node and returns a tuple containing a vector of new SDDP.Noise terms for the children of that node, and a vector of new SDDP.Noise terms for the stagewise-independent noise.

If f is called with the name of the root node (e.g., 0 in a linear policy graph, (0, 1) in a Markovian Policy Graph), then return a vector of SDDP.Noise for the children of the root node.

If use_insample_transition, the in-sample transition probabilities will be used. Therefore, f should only return a vector of the stagewise-independent noise terms, and f will not be called for the root node.

If terminate_on_cycle, terminate the forward pass once a cycle is detected. If max_depth > 0, return once max_depth nodes have been sampled. If terminate_on_dummy_leaf, terminate the forward pass with 1 - probability of sampling a child node.

Note that if terminate_on_cycle = false and terminate_on_dummy_leaf = false then max_depth must be set > 0.

You can use rollout_limit to set iteration specific depth limits. For example:

OutOfSampleMonteCarlo(rollout_limit = i -> 2 * i)

Example

# Given linear policy graph `graph` with `T` stages:
sampler = OutOfSampleMonteCarlo(graph) do node
    if node == 0
        return [SDDP.Noise(1, 1.0)]
    else
        noise_terms = [SDDP.Noise(node, 0.3), SDDP.Noise(node + 1, 0.7)]
        children = node < T ? [SDDP.Noise(node + 1, 0.9)] : SDDP.Noise{Int}[]
        return children, noise_terms
    end
end

# Given linear policy graph `graph` with `T` stages:
sampler = OutOfSampleMonteCarlo(graph, use_insample_transition=true) do node
    return [SDDP.Noise(node, 0.3), SDDP.Noise(node + 1, 0.7)]
end

Parallel schemes

SDDP.AbstractParallelScheme — Type

AbstractParallelScheme

Abstract type for different parallelism schemes.

SDDP.Serial — Type

Serial()

Run SDDP in serial mode.

SDDP.Asynchronous — Type

Asynchronous(init_callback::Function, slave_pids::Vector{Int} = workers())

Run SDDP in asynchronous mode workers with pid's slave_pids.

After initializing the models on each worker, call init_callback(model). Note that init_callback is run locally on the worker and not on the master thread.

Asynchronous(slave_pids::Vector{Int} = workers())

Run SDDP in asynchronous mode workers with pid's slave_pids.

Forward passes

SDDP.AbstractForwardPass — Type

AbstractForwardPass

Abstract type for different forward passes.

SDDP.DefaultForwardPass — Type

DefaultForwardPass()

The default forward pass.

SDDP.RevisitingForwardPass — Type

RevisitingForwardPass(
    period::Int = 500;
    sub_pass::AbstractForwardPass = DefaultForwardPass()
)

A forward pass scheme that generate period new forward passes (using sub_pass), then revisits all previously explored forward passes. This can be useful to encourage convergence at a diversity of points in the state-space.

Set period = typemax(Int) to disable.

For example, if period = 2, then the forward passes will be revisited as follows: 1, 2, 1, 2, 3, 4, 1, 2, 3, 4, 5, 6, 1, 2, ....

Risk Measures

SDDP.AbstractRiskMeasure — Type

AbstractRiskMeasure

The abstract type for the risk measure interface.

You need to define the following methods:

SDDP.adjust_probability

SDDP.adjust_probability — Function

adjust_probability(measure::Expectation
                   risk_adjusted_probability::Vector{Float64},
                   original_probability::Vector{Float64},
                   noise_support::Vector{Noise{T}},
                   objective_realizations::Vector{Float64},
                   is_minimization::Bool) where T

Integrality handlers

SDDP.AbstractIntegralityHandler — Type

AbstractIntegralityHandler

The abstract type for the integrality handlers interface.

SDDP.ContinuousRelaxation — Type

ContinuousRelaxation()

The continuous relaxation integrality handler. Duals are obtained in the backward pass by solving a continuous relaxation of each subproblem. Integrality constraints are retained in policy simulation.

SDDP.SDDiP — Type

SDDiP(; iteration_limit::Int = 100, atol::Float64, rtol::Float64)

The SDDiP integrality handler introduced by Zou, J., Ahmed, S. & Sun, X.A. Math. Program. (2019) 175: 461. Stochastic dual dynamic integer programming. https://doi.org/10.1007/s10107-018-1249-5.

Calculates duals by solving the Lagrangian dual for each subproblem. Kelley's method is used to compute Lagrange multipliers. iteration_limit controls the maximum number of iterations, and atol and rtol are the absolute and relative tolerances used in the termination criteria.

All state variables are assumed to take nonnegative values only.

Simulating the policy

SDDP.simulate — Function

simulate(
    model::PolicyGraph,
    number_replications::Int = 1,
    variables::Vector{Symbol} = Symbol[];
    sampling_scheme::AbstractSamplingScheme =
        InSampleMonteCarlo(),
    custom_recorders = Dict{Symbol, Function}(),
    require_duals::Bool = true,
    skip_undefined_variables::Bool = false,
    parallel_scheme::AbstractParallelScheme = Serial()
 )::Vector{Vector{Dict{Symbol, Any}}}

Perform a simulation of the policy model with number_replications replications using the sampling scheme sampling_scheme.

Returns a vector with one element for each replication. Each element is a vector with one-element for each node in the scenario that was sampled. Each element in that vector is a dictionary containing information about the subproblem that was solved.

In that dictionary there are four special keys:

:node_index, which records the index of the sampled node in the policy model

:noise_term, which records the noise observed at the node

:stage_objective, which records the stage-objective of the subproblem

:bellman_term, which records the cost/value-to-go of the node.

The sum of :stageobjective + :bellmanterm will equal the objective value of the solved subproblem.

In addition to the special keys, the dictionary will contain the result of JuMP.value(subproblem[key]) for each key in variables. This is useful to obtain the primal value of the state and control variables.

For more complicated data, the custom_recorders keyword argument can be used.

data = Dict{Symbol, Any}()
for (key, recorder) in custom_recorders
    data[key] = foo(subproblem)
end

For example, to record the dual of a constraint named my_constraint, pass the following:

simulation_results = SDDP.simulate(model, 2;
    custom_recorders = Dict{Symbol, Function}(
        :constraint_dual => (sp) -> JuMP.dual(sp[:my_constraint])
    )
)

The value of the dual in the first stage of the second replication can be accessed as:

simulation_results[2][1][:constraint_dual]

If you do not require dual variables (or if they are not available), pass require_duals = false.

If you attempt to simulate the value of a variable that is only defined in some of the stage problems, an error will be thrown. To over-ride this (and return a NaN instead), pass skip_undefined_variables = true.

Use parallel_scheme::[AbstractParallelScheme](@ref) to specify a scheme for simulating in parallel. Defaults to Serial.

SDDP.calculate_bound — Function

SDDP.calculate_bound(model::PolicyGraph, state::Dict{Symbol, Float64},
                       risk_measure=Expectation())

Calculate the lower bound (if minimizing, otherwise upper bound) of the problem model at the point state, assuming the risk measure at the root node is risk_measure.

SDDP.Historical — Type

Historical(scenarios::Vector{Vector{Tuple{T, S}}},
           probability::Vector{Float64})

A sampling scheme that samples a scenario from the vector of scenarios scenarios according to probability. If probability omitted, defaults to uniform probability.

Example

Historical(
    [
        [(1, 0.5), (2, 1.0), (3, 0.5)],
        [(1, 0.5), (2, 0.0), (3, 1.0)],
        [(1, 1.0), (2, 0.0), (3, 0.0)]
    ],
    [0.2, 0.5, 0.3]
)

Historical(scenario::Vector{Tuple{T, S}})

A deterministic sampling scheme that always samples scenario with probability 1.

Example

Historical([(1, 0.5), (2, 1.5), (3, 0.75)])

Decision rules

SDDP.DecisionRule — Type

DecisionRule(model::PolicyGraph{T}; node::T)

Create a decision rule for node node in model.

SDDP.evaluate — Function

evaluate(
    rule::DecisionRule;
    incoming_state::Dict{Symbol, Float64},
    noise = nothing,
    controls_to_record = Symbol[],
)

Evalute the decision rule rule at the point described by the incoming_state and noise.

If the node is deterministic, omit the noise argument.

Pass a list of symbols to controls_to_record to save the optimal primal solution corresponding to the names registered in the model.

evaluate(
    V::ValueFunction,
    point::Dict{Symbol, Float64},
    objective_state = nothing,
    belief_state = nothing
)

Evaluate the value function V at point in the state-space.

Returns a tuple containing the height of the function, and the subgradient w.r.t. the convex state-variables.

evalute(V::ValueFunction{Nothing, Nothing}; kwargs...)

Evalute the value function V at the point in the state-space specified by kwargs.

Example

evaluate(V; volume = 1)

evaluate(
    model::PolicyGraph{T}, test_scenarios::TestScenarios{T, S}
) where {T, S}

Evaluate the performance of the policy contained in model after a call to train on the scenarios specified by test_scenarios.

Example

model, test_scenarios = read_from_file("my_model.sof.json")
train(model; iteration_limit = 100)
simulations = evaluate(model, test_scenarios)

Visualizing the policy

SDDP.SpaghettiPlot — Type

SDDP.SpaghettiPlot(; stages, scenarios)

Initialize a new SpaghettiPlot with stages stages and scenarios number of replications.

SDDP.add_spaghetti — Function

SDDP.add_spaghetti(data_function::Function, plt::SpaghettiPlot; kwargs...)

Description

Add a new figure to the SpaghettiPlot plt, where the y-value of the scenarioth line when x = stage is given by data_function(plt.simulations[scenario][stage]).

Keyword arguments

xlabel: set the xaxis label

ylabel: set the yaxis label

title: set the title of the plot

ymin: set the minimum y value

ymax: set the maximum y value

cumulative: plot the additive accumulation of the value across the stages

interpolate: interpolation method for lines between stages.

Defaults to "linear" see the d3 docs for all options.

Examples

simulations = simulate(model, 10)
plt = SDDP.spaghetti_plot(simulations)
SDDP.add_spaghetti(plt; title = "Stage objective") do data
	return data[:stage_objective]
end

SDDP.publication_plot — Function

SDDP.publication_plot(
    data_function, simulations;
    quantile = [0.0, 0.1, 0.25, 0.5, 0.75, 0.9, 1.0],
    kwargs...)

Create a Plots.jl recipe plot of the simulations.

See Plots.jl for the list of keyword arguments.

Example

SDDP.publication_plot(simulations; title = "My title") do data
    return data[:stage_objective]
end

SDDP.ValueFunction — Type

ValueFunction

A representation of the value function. SDDP.jl uses the following unique representation of the value function that is undocumented in the literature.

It supports three types of state variables:

x - convex "resource" states

b - concave "belief" states

y - concave "objective" states

In addition, we have three types of cuts:

Single-cuts (also called "average" cuts in the literature), which involve the risk-adjusted expectation of the cost-to-go.

Multi-cuts, which use a different cost-to-go term for each realization w.

Risk-cuts, which correspond to the facets of the dual interpretation of a coherent risk measure.

Therefore, ValueFunction returns a JuMP model of the following form:

V(x, b, y) = min: μᵀb + νᵀy + θ
             s.t. # "Single" / "Average" cuts
                  μᵀb(j) + νᵀy(j) + θ >= α(j) + xᵀβ(j), ∀ j ∈ J
                  # "Multi" cuts
                  μᵀb(k) + νᵀy(k) + φ(w) >= α(k, w) + xᵀβ(k, w), ∀w ∈ Ω, k ∈ K
                  # "Risk-set" cuts
                  θ ≥ Σ{p(k, w) * φ(w)}_w - μᵀb(k) - νᵀy(k), ∀ k ∈ K

SDDP.evaluate — Method

evaluate(
    V::ValueFunction,
    point::Dict{Symbol, Float64},
    objective_state = nothing,
    belief_state = nothing
)

Evaluate the value function V at point in the state-space.

Returns a tuple containing the height of the function, and the subgradient w.r.t. the convex state-variables.

SDDP.plot — Function

plot(plt::SpaghettiPlot[, filename::String]; open::Bool = true)

The SpaghettiPlot plot plt to filename. If filename is not given, it will be saved to a temporary directory. If open = true, then a browser window will be opened to display the resulting HTML file.

Debugging the model

SDDP.write_subproblem_to_file — Function

write_subproblem_to_file(node::Node, filename::String; throw_error::Bool = false)

Write the subproblem contained in node to the file filename.

SDDP.deterministic_equivalent — Function

deterministic_equivalent(
    pg::PolicyGraph{T},
    optimizer = nothing;
    time_limit::Union{Real, Nothing} = 60.0
)

Form a JuMP model that represents the deterministic equivalent of the problem.

Examples

deterministic_equivalent(model)
deterministic_equivalent(model, GLPK.Optimizer)

StochOptFormat

SDDP.write_to_file — Function

write_to_file(
    model::PolicyGraph,
    filename::String;
    compression::MOI.FileFormats.AbstractCompressionScheme =
        MOI.FileFormats.AutomaticCompression(),
    kwargs...
)

Write model to filename in the StochOptFormat file format.

Pass an argument to compression to override the default of automatically detecting the file compression to use based on the extension of filename.

See Base.write(::IO, ::PolicyGraph) for information on the keyword arguments that can be provided.

WARNING: THIS FUNCTION IS EXPERIMENTAL. SEE THE FULL WARNING IN Base.write(::IO, ::PolicyGraph).

Example

write_to_file(model, "my_model.sof.json"; test_scenarios = 10)

SDDP.read_from_file — Function

read_from_file(
    filename::String;
    compression::MOI.FileFormats.AbstractCompressionScheme =
        MOI.FileFormats.AutomaticCompression(),
    kwargs...
)::Tuple{PolicyGraph, TestScenarios}

Return a tuple containing a PolicyGraph object and a TestScenarios read from filename in the StochOptFormat file format.

Pass an argument to compression to override the default of automatically detecting the file compression to use based on the extension of filename.

See Base.read(::IO, ::Type{PolicyGraph}) for information on the keyword arguments that can be provided.

WARNING: THIS FUNCTION IS EXPERIMENTAL. SEE THE FULL WARNING IN Base.read(::IO, ::Type{PolicyGraph}).

Example

model, test_scenarios = read_from_file("my_model.sof.json")

Base.write — Method

Base.write(
    io::IO,
    model::PolicyGraph;
    test_scenarios::Union{Int, TestScenarios} = 1_000,
    kwargs...
)

Write model to io in the StochOptFormat file format.

Pass an Int to test_scenarios (default 1_000) to specify the number of test scenarios to generate using the InSampleMonteCarlo sampling scheme. Alternatively, pass a TestScenarios object to manually specify the test scenarios to use.

Any additional kwargs passed to write will be stored in the top-level of the resulting StochOptFormat file. Valid arguments include name, author, date, and description.

WARNING: THIS FUNCTION IS EXPERIMENTAL. THINGS MAY CHANGE BETWEEN COMMITS. YOU SHOULD NOT RELY ON THIS FUNCTIONALITY AS A LONG-TERM FILE FORMAT (YET).

In addition to potential changes to the underlying format, only a subset of possible modifications are supported. These include:

JuMP.fix

JuMP.set_lower_bound

JuMP.set_upper_bound

JuMP.set_normalized_rhs

Changes to the constant or affine terms in a stage objective

If your model uses something other than this, this function will silently write an incorrect formulation of the problem.

Example

open("my_model.sof.json", "w") do io
    write(
        io,
        model;
        test_scenarios = 10,
        name = "MyModel",
        author = "@odow",
        date = "2020-07-20",
        description = "Example problem for the SDDP.jl documentation",
    )
end

Base.read — Method

Base.read(
    io::IO,
    ::Type{PolicyGraph};
    bound::Float64 = 1e6,
)::Tuple{PolicyGraph, TestScenarios}

Return a tuple containing a PolicyGraph object and a TestScenarios read from io in the StochOptFormat file format.

See also: evaluate.

WARNING: THIS FUNCTION IS EXPERIMENTAL. THINGS MAY CHANGE BETWEEN COMMITS. YOU SHOULD NOT RELY ON THIS FUNCTIONALITY AS A LONG-TERM FILE FORMAT (YET).

In addition to potential changes to the underlying format, only a subset of possible modifications are supported. These include:

Additive random variables in the constraints or in the objective

Multiplicative random variables in the objective

If your model uses something other than this, this function may throw an error or silently build a non-convex model.

Example

open("my_model.sof.json", "r") do io
    model, test_scenarios = read(io, PolicyGraph)
end

SDDP.evaluate — Method

evaluate(
    model::PolicyGraph{T}, test_scenarios::TestScenarios{T, S}
) where {T, S}

Evaluate the performance of the policy contained in model after a call to train on the scenarios specified by test_scenarios.

Example

model, test_scenarios = read_from_file("my_model.sof.json")
train(model; iteration_limit = 100)
simulations = evaluate(model, test_scenarios)

SDDP.TestScenarios — Type

TestScenarios{T, S}(scenarios::Vector{TestScenario{T, S}})

An AbstractSamplingScheme based on a vector of scenarios.

Each scenario is a vector of Tuple{T, S} where the first element is the node to visit and the second element is the realization of the stagewise-independent noise term. Pass nothing if the node is deterministic.

SDDP.TestScenario — Type

TestScenario{T, S}(probability::Float64, scenario::Vector{Tuple{T, S}})

A single scenario for testing.

See also: TestScenarios.