Reference

API Reference

Policy graphs

SDDP.Graph — Type.

Graph(root_node::T) where T

Create an empty graph struture with the root node root_node.

SDDP.add_node — Function.

add_node(graph::Graph{T}, node::T) where T

Add a node to the graph graph.

Examples

add_node(graph, :A)

SDDP.add_edge — Function.

add_edge(graph::Graph{T}, edge::Pair{T, T}, probability::Float64) where T

Add an edge to the graph graph.

Examples

add_edge(graph, 1 => 2, 0.9)
add_edge(graph, :root => :A, 1.0)

SDDP.add_ambiguity_set — Function.

add_ambiguity_set(graph::Graph{T}, set::Vector{T})

Add set to the belief partition of graph.

Examples

graph = LinearGraph(3)
add_ambiguity_set(graph, [1, 2])
add_ambiguity_set(graph, [3])

SDDP.LinearGraph — Function.

LinearGraph(stages::Int)

SDDP.MarkovianGraph — Function.

MarkovianGraph(transition_matrices::Vector{Matrix{Float64}})

MarkovianGraph(; stages::Int,
               transition_matrix::Matrix{Float64},
               root_node_transition::Vector{Float64})

Construct a Markovian graph object.

SDDP.LinearPolicyGraph — Function.

LinearPolicyGraph(builder::Function; stages::Int, kwargs...)

Create a linear policy graph with stages number of stages.

See SDDP.PolicyGraph for the other keyword arguments.

SDDP.MarkovianPolicyGraph — Function.

MarkovianPolicyGraph(builder::Function;
    transition_matrices::Vector{Array{Float64, 2}}, kwargs...)

Create a Markovian policy graph based on the transition matrices given in transition_matrices.

See SDDP.PolicyGraph for the other keyword arguments.

SDDP.PolicyGraph — Type.

PolicyGraph(builder::Function, graph::Graph{T};
    sense = :Min,
    bellman_function = nothing,
    lower_bound = -Inf,
    upper_bound = Inf,
    optimizer = nothing,
    direct_mode = true,
    lipschitz_belief = Dict{T, Float64}()) where {T}

Construct a policy graph based on the graph structure of graph. (See SDDP.Graph for details.)

Example

function builder(subproblem::JuMP.Model, index)
    # ... subproblem definition ...
end

model = PolicyGraph(builder, graph;
                    lower_bound = 0.0,
                    optimizer = with_optimizer(GLPK.Optimizer),
                    direct_mode = false)

Or, using the Julia do ... end syntax:

model = PolicyGraph(graph;
                    lower_bound = 0.0,
                    optimizer = with_optimizer(GLPK.Optimizer),
                    direct_mode = true) do subproblem, index
    # ... subproblem definitions ...
end

Subproblem definition

SDDP.@stageobjective — Macro.

@stageobjective(subproblem, expr)

Set the stage-objective of subproblem to expr.

Example

@stageobjective(subproblem, 2x + y)

SDDP.parameterize — Function.

parameterize(modify::Function,
             subproblem::JuMP.Model,
             realizations::Vector{T},
             probability::Vector{Float64} = fill(1.0 / length(realizations))
                 ) where T

Add a parameterization function modify to subproblem. The modify function takes one argument and modifies subproblem based on the realization of the noise sampled from realizations with corresponding probabilities probability.

In order to conduct an out-of-sample simulation, modify should accept arguments that are not in realizations (but still of type T).

Example

SDDP.parameterize(subproblem, [1, 2, 3], [0.4, 0.3, 0.3]) do ω
    JuMP.set_upper_bound(x, ω)
end

parameterize(node::Node, noise)

Parameterize node node with the noise noise.

SDDP.add_objective_state — Function.

add_objective_state(update::Function, subproblem::JuMP.Model; kwargs...)

Add an objective state variable to subproblem.

Required kwargs are:

initial_value: The initial value of the objective state variable at the root node.

lipschitz: The lipschitz constant of the objective state variable.

Setting a tight value for the lipschitz constant can significantly improve the speed of convergence.

Optional kwargs are:

lower_bound: A valid lower bound for the objective state variable. Can be -Inf.

upper_bound: A valid upper bound for the objective state variable. Can be +Inf.

Setting tight values for these optional variables can significantly improve the speed of convergence.

If the objective state is N-dimensional, each keyword argument must be an NTuple{N, Float64}. For example, initial_value = (0.0, 1.0).

SDDP.objective_state — Function.

objective_state(subproblem::JuMP.Model)

Return the current objective state of the problem.

Can only be called from SDDP.parameterize.

Training the policy

SDDP.numerical_stability_report — Function.

numerical_stability_report([io::IO=stdout,] model::PolicyGraph,
                           by_node::Bool=false, print=true, warn::Bool=true)

Print a report identifying possible numeric stability issues.

If by_node, print a report for each node in the graph.

If print, print to io.

If warn, warn if the coefficients may cause numerical issues.

SDDP.train — Function.

SDDP.train(model::PolicyGraph; kwargs...)

Train the policy for model. Keyword arguments:

iteration_limit::Int: number of iterations to conduct before termination

time_limit::Float64: number of seconds to train before termination

stoping_rules: a vector of SDDP.AbstractStoppingRule

print_level: control the level of printing to the screen

log_file: filepath at which to write a log of the training progress

runnumericalstability_report: generate a numerical stability report prior to solve

refine_at_similar_nodes::Bool: if SDDP can detect that two nodes have the same children, it can cheaply add a cut discovered at one to the other. In almost all cases this should be set to true.

cut_deletion_minimum::Int: the minimum number of cuts to cache before deleting cuts from the subproblem. This is solver specific; however, smaller values result in smaller subproblems, at the expense of more time spent performing cut selection.

risk_measure: the risk measure to use at each node.

sampling_scheme: a sampling scheme to use on the forward pass of the algorithm. Defaults to InSampleMonteCarlo().

There is also a special option for infinite horizon problems

cyclediscretizationdelta: the maximum distance between states allowed on the forward pass. This is for advanced users only and needs to be used in conjunction with a different sampling_scheme.

SDDP.termination_status — Function.

termination_status(model::PolicyGraph)

Query the reason why the training stopped.

SDDP.write_cuts_to_file — Function.

write_cuts_to_file(model::PolicyGraph{T}, filename::String) where {T}

Write the cuts that form the policy in model to filename in JSON format.

See also SDDP.read_cuts_from_file.

SDDP.read_cuts_from_file — Function.

read_cuts_from_file(
    model::PolicyGraph{T}, filename::String;
    node_name_parser::Function = _node_name_parser) where {T}

Read cuts (saved using SDDP.write_cuts_to_file) from filename into model.

Since T can be an arbitrary Julia type, the conversion to JSON is lossy. When reading, read_cuts_from_file only supports T=Int, T=NTuple{N, Int}, and T=Symbol. If you have manually created a policy graph with a different node type T, provide a function node_name_parser with the signature node_name_parser(T, name::String)::T where {T} that returns the name of each node given the string name name.

See also SDDP.write_cuts_to_file.

Stopping rules

SDDP.AbstractStoppingRule — Type.

AbstractStoppingRule

The abstract type for the stopping-rule interface.

You need to define the following methods:

SDDP.stopping_rule_status

SDDP.convergence_test

SDDP.stopping_rule_status — Function.

stopping_rule_status(::AbstractStoppingRule)::Symbol

Return a symbol describing the stopping rule.

SDDP.convergence_test — Function.

convergence_test(model::PolicyGraph, log::Vector{Log}, ::AbstractStoppingRule)::Bool

Return a Bool indicating if the algorithm should terminate the training.

Sampling schemes

SDDP.AbstractSamplingScheme — Type.

AbstractSamplingScheme

The abstract type for the sampling-scheme interface.

You need to define the following methods:

SDDP.sample_scenario

SDDP.sample_scenario — Function.

sample_scenario(graph::PolicyGraph{T}, ::AbstractSamplingScheme) where T

Sample a scenario from the policy graph graph based on the sampling scheme.

Returns ::Tuple{Vector{Tuple{T, <:Any}}, Bool}, where the first element is the scenario, and the second element is a Boolean flag indicating if the scenario was terminated due to the detection of a cycle.

The scenario is a list of tuples (type Vector{Tuple{T, <:Any}}) where the first component of each tuple is the index of the node, and the second component is the stagewise-independent noise term observed in that node.

Sampling schemes

SDDP.AbstractRiskMeasure — Type.

AbstractRiskMeasure

The abstract type for the risk measure interface.

You need to define the following methods:

SDDP.adjust_probability

SDDP.adjust_probability — Function.

adjust_probability(measure::Expectation
                   risk_adjusted_probability::Vector{Float64},
                   original_probability::Vector{Float64},
                   noise_support::Vector{Noise{T}},
                   objective_realizations::Vector{Float64},
                   is_minimization::Bool) where T

Simulating the policy

SDDP.simulate — Function.

simulate(model::PolicyGraph,
         number_replications::Int = 1,
         variables::Vector{Symbol} = Symbol[];
         sampling_scheme::AbstractSamplingScheme =
             InSampleMonteCarlo(),
         custom_recorders = Dict{Symbol, Function}()
 )::Vector{Vector{Dict{Symbol, Any}}}

Perform a simulation of the policy model with number_replications replications using the sampling scheme sampling_scheme.

Returns a vector with one element for each replication. Each element is a vector with one-element for each node in the scenario that was sampled. Each element in that vector is a dictionary containing information about the subproblem that was solved.

In that dictionary there are four special keys:

:node_index, which records the index of the sampled node in the policy model

:noise_term, which records the noise observed at the node

:stage_objective, which records the stage-objective of the subproblem

:bellman_term, which records the cost/value-to-go of the node.

The sum of :stageobjective + :bellmanterm will equal the objective value of the solved subproblem.

In addition to the special keys, the dictionary will contain the result of JuMP.value(subproblem[key]) for each key in variables. This is useful to obtain the primal value of the state and control variables.

For more complicated data, the custom_recorders keyword arguement can be used.

data = Dict{Symbol, Any}()
for (key, recorder) in custom_recorders
    data[key] = foo(subproblem)
end

For example, to record the dual of a constraint named my_constraint, pass the following:

simulation_results = simulate(model, number_replications=2;
    custom_recorders = Dict(
        :constraint_dual = (sp) -> JuMP.dual(sp[:my_constraint])
    )
)

The value of the dual in the first stage of the second replication can be accessed as:

simulation_results[2][1][:constraint_dual]

SDDP.calculate_bound — Function.

SDDP.calculate_bound(model::PolicyGraph, state::Dict{Symbol, Float64},
                       risk_measure=Expectation())

Calculate the lower bound (if minimizing, otherwise upper bound) of the problem model at the point state, assuming the risk measure at the root node is risk_measure.

SDDP.Historical — Type.

Historical(scenarios::Vector{Vector{Tuple{T, S}}},
           probability::Vector{Float64})

A sampling scheme that samples a scenario from the vector of scenarios scenarios according to probability. If probability omitted, defaults to uniform probability.

Example

Historical(
    [
        [(1, 0.5), (2, 1.0), (3, 0.5)],
        [(1, 0.5), (2, 0.0), (3, 1.0)],
        [(1, 1.0), (2, 0.0), (3, 0.0)]
    ],
    [0.2, 0.5, 0.3]
)

Historical(scenario::Vector{Tuple{T, S}})

A deterministic sampling scheme that always samples scenario with probability 1.

Example

Historical([(1, 0.5), (2, 1.5), (3, 0.75)])

Visualizing the policy

SDDP.SpaghettiPlot — Type.

SDDP.SpaghettiPlot(; stages, scenarios)

Initialize a new SpaghettiPlot with stages stages and scenarios number of replications.

SDDP.add_spaghetti — Function.

SDDP.add_spaghetti(data_function::Function, plt::SpaghettiPlot; kwargs...)

Description

Add a new figure to the SpaghettiPlot plt, where the y-value of the scenarioth line when x = stage is given by data_function(plt.simulations[scenario][stage]).

Keyword arguments

xlabel: set the xaxis label

ylabel: set the yaxis label

title: set the title of the plot

ymin: set the minimum y value

ymax: set the maximum y value

cumulative: plot the additive accumulation of the value across the stages

interpolate: interpolation method for lines between stages.

Defaults to "linear" see the d3 docs for all options.

Examples

simulations = simulate(model, 10)
plt = SDDP.spaghetti_plot(simulations)
SDDP.add_spaghetti(plt; title = "Stage objective") do data
	return data[:stage_objective]
end

SDDP.publication_plot — Function.

SDDP.publication_plot(
    data_function, simulations;
    quantile = [0.0, 0.1, 0.25, 0.5, 0.75, 0.9, 1.0],
    kwargs...)

Create a Plots.jl recipe plot of the simulations.

See Plots.jl for the list of keyword arguments.

Example

SDDP.publication_plot(simulations; title = "My title") do data
    return data[:stage_objective]
end

Debugging the model

SDDP.write_subproblem_to_file — Function.

write_subproblem_to_file(node::Node, filename::String; format=:both)

Write the subproblem contained in node to the file filename.

format should be one of :mps, :lp, or :both.