Reference

API Reference

Defining the model

<!– SDDPModel –>

States

SDDP.@state — Macro.

@state(sp, stateleaving, stateentering)

Description

Define a new state variable in the subproblem sp.

Arguments

sp the subproblem

stateleaving any valid JuMP @variable syntax to define the value of the state variable at the end of the stage

stateentering any valid JuMP @variable syntax to define the value of the state variable at the beginning of the stage

Examples

@state(sp, 0 <= x[i=1:3] <= 1, x0==rand(3)[i] )

@state(sp, y <= 1, y0==0.5 )

@state(sp, z , z0==0.5 )

source

SDDP.@states — Macro.

@states(sp, begin
    stateleaving1, stateentering1
    stateleaving2, stateentering2
end)

Description

Define a new state variables in the subproblem sp.

Arguments

sp the subproblem

stateleaving any valid JuMP @variable syntax to define the value of the state variable at the end of the stage

stateentering any valid JuMP @variable syntax to define the value of the state variable at the beginning of the stage

Usage

@states(sp, begin
    0 <= x[i=1:3] <= 1, x0==rand(3)[i]
         y        <= 1, y0==0.5
         z            , z0==0.5
 end)

source

Noises

SDDP.@noise — Macro.

@noise(sp, rhs, constraint)

Description

Add a constraint with a noise in the RHS vector to the subproblem sp.

Arguments

sp the subproblem

rhs keyword argument key=value where value is a one-dimensional array containing the noise realisations

constraint any valid JuMP @constraint syntax that includes the keyword defined by rhs

Examples

@noise(sp, i=1:2, x + y <= i )

@noise(sp, i=1:2, x + y <= 3 * rand(2)[i] )

source

SDDP.@noises — Macro.

@noises(sp, rhs, begin
    constraint
end)

Description

The plural form of @noise similar to the JuMP macro @constraints.

Arguments

See @noise.

Examples

@noises(sp, i=1:2, begin
    x + y <= i
    x + y <= 3 * rand(2)[i]
end)

source

SDDP.setnoiseprobability! — Function.

setnoiseprobability!(sp::JuMP.Model, distribution::Vector{Float64})

Description

Set the probability distribution of the stagewise independent noise in the sp subproblem.

Arguments

sp the subproblem

distribution vector containing the probability of each outcome occuring. Should sum to 1. Defaults to the uniform distribution.

Examples

If there are two realizations:

setnoiseprobability!(sp, [0.3, 0.7])

setnoiseprobability!(sp, [0.5, 0.6]) will error!

source

Objective

<!– stageobjective! –>

Risk Measures

SDDP.AbstractRiskMeasure — Type.

AbstractRiskMeasure

Description

Abstract type for all risk measures.

source

SDDP.Expectation — Type.

Expectation()

Description

The expectation risk measure.

source

SDDP.NestedAVaR — Type.

NestedAVaR(;lambda=1.0, beta=1.0)

Description

A risk measure that is a convex combination of Expectation and Average Value @ Risk (also called Conditional Value @ Risk).

λ * E[x] + (1 - λ) * AV@R(1-β)[x]

Keyword Arguments

lambda

Convex weight on the expectation ((1-lambda) weight is put on the AV@R component. Inreasing values of lambda are less risk averse (more weight on expecattion)

beta

The quantile at which to calculate the Average Value @ Risk. Increasing values of beta are less risk averse. If beta=0, then the AV@R component is the worst case risk measure.

Returns

`m::NestedAVaR<:AbstractRiskMeasure`

source

SDDP.modifyprobability! — Function.

modifyprobability!(measure::AbstractRiskMeasure,
        riskadjusted_distribution,
        original_distribution::Vector{Float64},
        observations::Vector{Float64},
        m::SDDPModel,
        sp::JuMP.Model
)

Description

Calculate the risk-adjusted probability of each scenario using the 'change-of-probabilities' approach of Philpott, de Matos, and Finardi,(2013). On solving multistage stochastic programs with coherent risk measures. Operations Research 61(4), 957-970.

Arguments

measure::AbstractRiskMeasure

The risk measure

riskadjusted_distribution

A new probability distribution

original_distribution::Vector{Float64}

The original probability distribution.

observations::Vector{Float64}

The vector of objective values from the next stage problems (one for each scenario).

m::SDDPModel

The full SDDP model

sp::JuMP.Model

The stage problem that the cut will be added to.

source

Cut Oracles

SDDP.AbstractCutOracle — Type.

AbstractCutOracle

Description

Abstract type for all cut oracles.

source

SDDP.DefaultCutOracle — Type.

DefaultCutOracle()

Description

Initialize the default cut oracle.

This oracle keeps every cut discovered and does not perform cut selection.

source

SDDP.LevelOneCutOracle — Function.

LevelOneCutOracle()

Description

Initialize the cut oracle for Level One cut selection. See:

V. de Matos,A. Philpott, E. Finardi, Improving the performance of Stochastic Dual Dynamic Programming, Journal of Computational and Applied Mathematics 290 (2015) 196–208.

source

SDDP.storecut! — Function.

storecut!(oracle::AbstactCutOracle, m::SDDPModel, sp::JuMP.Model, cut::Cut)

Description

Store the cut cut in the Cut Oracle oracle. oracle will belong to the subproblem sp in the SDDPModel m.

source

SDDP.validcuts — Function.

validcuts(oracle::AbstactCutOracle)

Description

Return an iterable list of all the valid cuts contained within oracle.

source

Solve

<!– SDDP.solve –>

SDDP.MonteCarloSimulation — Type.

MonteCarloSimulation(;kwargs...)

Description

Collection of settings to control the simulation phase of the SDDP solution process.

Arguments

frequency::Int

The frequency (by iteration) with which to run the policy simulation phase of the algorithm in order to construct a statistical bound for the policy. Defaults to 0 (never run).

min::Float64

Minimum number of simulations to conduct before constructing a confidence interval for the bound. Defaults to 20.

step::Float64

Number of additional simulations to conduct before constructing a new confidence interval for the bound. Defaults to 1.

max::Float64

Maximum number of simulations to conduct in the policy simulation phase. Defaults to min.

confidence::Float64

Confidence level of the confidence interval. Defaults to 0.95 (95% CI).

termination::Bool

Whether to terminate the solution algorithm with the status :converged if the deterministic bound is with in the statistical bound after max simulations. Defaults to false.

source

SDDP.BoundConvergence — Type.

BoundConvergence(;kwargs...)

Description

Collection of settings to control the bound stalling convergence test.

Arguments

iterations::Int

Terminate if the maximum deviation in the deterministic bound from the mean over the last iterations number of iterations is less than rtol (in relative terms) or atol (in absolute terms).

rtol::Float64

Maximum allowed relative deviation from the mean. Defaults to 0.0

atol::Float64

Maximum allowed absolute deviation from the mean. Defaults to 0.0

source

Results

<!– simulate –> <!– @visualise –>

SDDP.getbound — Function.

getbound(m)

Description

Get the lower (if minimizing), or upper (if maximizing) bound of the solved SDDP model m.

source

Save/Load

<!– loadcuts! –>