Quick Start

Quick Start

Initialising the model object

The first step is to initialise the SDDP model object. We do this using the following syntax:

If we have more than one markov state:

m = SDDPModel([;kwargs...]) do sp, stage, markov_state

    # Stage problem definition where `sp` is a `JuMP.Model`object,

end

Otherwise if we have a single markov state

m = SDDPModel([;kwargs...]) do sp, stage

  # Stage problem definition

end

SDDP.SDDPModel — Type.

SDDPModel(;kwargs...) do ...

end

Description

This function constructs an SDDPModel.

Required Keyword arguments

stages::Int

The number of stages in the problem. A stage is defined as each step in time at which a decion can be made. Defaults to 1.

objective_bound::Float64

solver::MathProgBase.AbstractMathProgSolver

Optional Keyword arguments

cut_oracle

risk_measure

noise_probability

markov_transition

Returns

* `m`: the `SDDPModel`

source

Solve

To solve the SDDP model m we use status = solve(m::SDDPModel; kwargs...). This accepts a number of keyword arguments to control the solution process.

JuMP.solve — Function.

solve(m::SDDPModel; kwargs...)

Description

Solve the SDDPModel m using SDDP. Accepts a number of keyword arguments to control the solution process.

Positional arguments

m: the SDDPModel to solve

Keyword arguments

max_iterations::Int: The maximum number of cuts to add to a single stage problem before terminating. Defaults to 10.

time_limit::Real: The maximum number of seconds (in real time) to compute for before termination. Defaults to Inf.

simulation::MonteCarloSimulation: We control the behaviour of the policy simulation phase of the algorithm using the MonteCarloSimulation(;kwargs...) constructor. This just groups a series of related keyword arguments. The keywords are

frequency::Int

The frequency (by iteration) with which to run the policy simulation phase of the algorithm in order to construct a statistical bound for the policy. Defaults to 0 (never run).

min::Float64

Minimum number of simulations to conduct before constructing a confidence interval for the bound. Defaults to 20.

step::Float64

Number of additional simulations to conduct before constructing a new confidence interval for the bound. Defaults to 1.

max::Float64

Maximum number of simulations to conduct in the policy simulation phase. Defaults to min.

confidence::Float64

Confidence level of the confidence interval. Defaults to 0.95 (95% CI).

termination::Bool

Whether to terminate the solution algorithm with the status :converged if the deterministic bound is with in the statistical bound after max simulations. Defaults to false.

bound_convergence: We may also wish to terminate the algorithm if the deterministic bound stalls for a specified number of iterations (regardless of whether the policy has converged). This can be controlled by the BoundConvergence(;kwargs...) constructor. It has the following keywords:

iterations::Int

Terminate if the maximum deviation in the deterministic bound from the mean over the last iterations number of iterations is less than rtol (in relative terms) or atol (in absolute terms).

rtol::Float64

Maximum allowed relative deviation from the mean. Defaults to 0.0

atol::Float64

Maximum allowed absolute deviation from the mean. Defaults to 0.0

cut_selection_frequency::Int: Frequency (by iteration) with which to rebuild subproblems using a subset of cuts. Frequent cut selection (i.e. cut_selection_frequency is small) reduces the size of the subproblems that are solved, but incurrs the overhead of rebuilding the subproblems. However, infrequent cut selection (i.e. cut_selection_frequency is large) allows the subproblems to grow large (many constraints) leading to an increase in the solve time of individual subproblems. Defaults to 0 (never run).

print_level::Int: 0 - off: nothing logged to screen (still written to log file if specified). 1 - on: solve iterations written to screen. Defaults to 1

log_file::String: Relative filename to write the log to disk. Defaults to "" (no log written)

solve_type: One of

Asyncronous() - solve using a parallelised algorithm

Serial() - solve using a serial algorithm

Default chooses automatically based on the number of available processors.

reduce_memory_footprint::Bool: Implements the idea proposed in https://github.com/JuliaOpt/JuMP.jl/issues/969#issuecomment-282191105 to reduce the memory consumption when running SDDP. This is an issue if you wish to save the model m to disk since it discards important information. Defaults to false.

cut_output_file::String: Relative filename to write discovered cuts to disk. Defaults to "" (no cuts written)

Returns

status::Symbol: Reason for termination. One of

:solving

:interrupted

:converged

:max_iterations

:bound_convergence

:time_limit

source